+++ app/src/App/index.test.tsx
@@ -1,6 +1,6 @@
 import "react-native";
 import React from "react";
-import App from "./App";
+import App from "./index";
 
 // Note: test renderer must be required after react-native.
 import renderer from "react-test-renderer";
